# docs/openapi.yml
openapi: 3.0.3
info:
  title: RIMAC Medical Appointment API
  description: |
    Sistema serverless completo de gestión de citas médicas para RIMAC.
    
    ## Características
    - ✅ Procesamiento asíncrono por país (PE/CL)
    - ✅ Actualizaciones en tiempo real con EventBridge
    - ✅ Panel de administración con JWT
    - ✅ Arquitectura event-driven escalable
    - ✅ Base de datos dual: DynamoDB + RDS MySQL
    - ✅ Dashboard con estadísticas en tiempo real
    
    ## Flujo del Sistema
    1. **Cliente** crea cita → DynamoDB (pending)
    2. **SNS/SQS** → Procesamiento por país
    3. **Lambda específico** → RDS save (pending)
    4. **EventBridge** → Confirmación automática
    5. **Status update** → RDS (completed)
    6. **Admin Dashboard** → Estadísticas en tiempo real
    
  version: 2.0.0
  contact:
    name: RIMAC API Support
    email: api-support@rimac.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://bnh6zvyh59.execute-api.us-east-2.amazonaws.com/dev
    description: Desarrollo
  - url: https://api-prod.rimac.com
    description: Producción
...
paths:
...
  /appointments:
    post:
      summary: Crear nueva cita médica
      description: |
        Crea una nueva cita médica y la procesa asíncronamente según el país.
        
        ## Flujo de Procesamiento
        1. Validación de datos de entrada
        2. Creación en DynamoDB con status "pending"
        3. Envío a cola SNS/SQS por país
        4. Procesamiento por Lambda específico (PE/CL)
        5. Guardado en RDS del país correspondiente
        6. Confirmación vía EventBridge
        7. Actualización de status a "completed"
        
      operationId: createAppointment
      tags:
        - Appointments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
            examples:
              peru:
                summary: Cita en Perú
                value:
                  insuredId: "12345"
                  scheduleId: 100
                  countryISO: "PE"
              chile:
                summary: Cita en Chile
                value:
                  insuredId: "67890"
                  scheduleId: 200
                  countryISO: "CL"
      responses:
        '201':
          description: Cita creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAppointmentResponse'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /appointments/{insuredId}:
    get:
      summary: Obtener citas de un asegurado
      description: |
        Obtiene todas las citas médicas de un asegurado específico desde DynamoDB.
        
        Las citas se ordenan por fecha de creación (más recientes primero).
      operationId: getAppointmentsByInsuredId
      tags:
        - Appointments
      parameters:
        - name: insuredId
          in: path
          required: true
          description: ID del asegurado (5 dígitos)
          schema:
            type: string
            pattern: '^[0-9]{5}$'
            example: "12345"
      responses:
        '200':
          description: Citas encontradas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAppointmentsResponse'
        '404':
          description: No se encontraron citas para el asegurado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============= ADMIN API =============
...
components:
  securitySchemes:
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtenido desde `/admin/auth`.
        
        **Formato:** `Bearer <token>`
  schemas:
...
    CreateAppointmentRequest:
      type: object
      required:
        - insuredId
        - scheduleId
        - countryISO
      properties:
        insuredId:
          type: string
          pattern: '^[0-9]{5}$'
          description: ID del asegurado (exactamente 5 dígitos)
          example: "12345"
        scheduleId:
          type: integer
          minimum: 1
          maximum: 99999
          description: ID del horario disponible
          example: 100
        countryISO:
          type: string
          enum: [PE, CL]
          description: Código ISO del país
          example: "PE"
    CreateAppointmentResponse:
      type: object
      properties:
        message:
          type: string
          example: "Appointment request is being processed"
        appointmentId:
          type: string
          format: uuid
          description: ID único de la cita generado
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [pending, completed, cancelled]
          example: "pending"
        countryISO:
          type: string
          enum: [PE, CL]
          example: "PE"
        createdAt:
          type: string
          format: date-time
          example: "2025-09-13T22:30:00.000Z"
    GetAppointmentsResponse:
      type: object
      properties:
        insuredId:
          type: string
          example: "12345"
        count:
          type: integer
          description: Número total de citas encontradas
          example: 3
        appointments:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
    Appointment:
      type: object
      properties:
        appointmentId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        insuredId:
          type: string
          example: "12345"
        scheduleId:
          type: integer
          example: 100
        countryISO:
          type: string
          enum: [PE, CL]
          example: "PE"
        status:
          type: string
          enum: [pending, completed, cancelled]
          example: "completed"
        createdAt:
          type: string
          format: date-time
          example: "2025-09-13T22:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-09-13T22:30:15.000Z"

    # ========== ADMIN SCHEMAS ==========
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Descripción del error
          example: "Invalid request parameters"
        message:
          type: string
          description: Mensaje detallado del error
          example: "insuredId must be exactly 5 digits"
        details:
          type: object
          description: Detalles adicionales del error (opcional)

# ============= EJEMPLO =============
...